requirements.txt
asttokens==2.4.0
backcall==0.2.0
comm==0.1.4
debugpy==1.7.0
decorator==5.1.1
delta-spark==2.4.0
exceptiongroup==1.1.3
executing==1.2.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.2
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
plotly==5.16.1
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==13.0.0
Pygments==2.16.1
pyspark==3.4.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
six==1.16.0
stack-data==0.6.2
tenacity==8.2.3
tomli==2.0.1
tornado==6.3.3
traitlets==5.9.0
tzdata==2023.3
wcwidth==0.2.6
zipp==3.16.2
